create custom elasticsearchconverter for passport retrieval
There was a problem with ES document to entity class mapping during searches:
the `ElasticsearchOperations` bean used for that was unable to handle document passports,
so that passports of search results would always be empty, even though entity lists queried using
their `AbstractElasticsearchRepository`s had correctly mapped passports.

The solution to this problem was to pass the `ElasticsearchRestTemplate` constructor
an `ElasticsearchConverter` instance during during creation of the `ElasticsearchRestTemplate`
(which gets injected when autowiring an `ElasticsearchOperations` bean).
To this end just copy `AbstractElasticsearchTemplate.createElasticsearchConverter`, and
insert a `setConversions` call passing our `ElasticsearchCustomConversions` bean.