move native query creation to own package
take all the query building stuff from the lemma service and put it into its own
class which does nothing but create native search query parts (query, aggregations,
sort settings...) which get then put together by the abstract superclass.

The idea is to have various entity query builder factories designed for specific types
of search command and requested entity type, and have services know about them so that
they can look up the ones appropriate for incoming requests.

A specialized subclass of the new `AbstractEntityQueryBuilder` is meant for scenarios
where instead of a page of retrieved entities, it might be necessary to get hold of
only the IDs of all entities of a certain type matching a search command's criteria.
Things like this are going to be needed for the more complex of the proposed search modes,
where query expansion and outer joins will require multiple combined searches in one
request. For instance, for searching lemma occurrences in texts from a given time period
and a geographic area, we would first query all thesaurus entries who are descendents
from the two `findSpot` and `date` entries put in the search form, as the concepts
represented by those are just more specific sub-categories of the same concept.
So we need to expand the original query with the two thesaurus terms by all of their
subordinated terms. We then take all of their IDs and use them in a `terms` query
on the text object index in order to narrow down the candidates for lemma occurrence
search, and so on.