rm index exists test
ok so for some reason the order in which github CI executes the junit tests has changed, and now
first test class being run is `ConnectionTest`. Because at this point no spring application context
has yet been booted up, spring data hasn't created the elasticsearch indices yet, causing a test
checking the `lemma` index to fail. We can fix this by either making `ConnectionTest` a
`@SpringBootTest` so that it causes the `lemma` index to be created itself (thusly slowing it down),
or by removing this test as it is of limited use anyway.